requirements.txt
Flask==3.0.3
pymongo==4.6.1
python-dotenv==1.0.1
flask-cors==4.0.0
workos==5.15.0
